# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /books:
    get:
      operationId: listBooks
      tags: [books]
      summary: List books
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: categoryId
          in: query
          schema:
            type: integer
        - name: rackId
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of books
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookListResponse"
    post:
      operationId: createBook
      tags: [books]
      summary: Create a book
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Book"
  /books/public:
    get:
      operationId: listPublicBooks
      tags: [books]
      summary: List books publicly
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: categoryId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of books
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookListResponse"
  /books/{id}:
    patch:
      operationId: updateBook
      tags: [books]
      summary: Update a book
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Book"
    delete:
      operationId: deleteBook
      tags: [books]
      summary: Delete a book
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List categories
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [categories]
      summary: Create a category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
  /categories/{id}:
    patch:
      operationId: updateCategory
      tags: [categories]
      summary: Update a category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
    delete:
      operationId: deleteCategory
      tags: [categories]
      summary: Delete a category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
  /borrowings:
    get:
      operationId: listBorrowings
      tags: [borrowings]
      summary: List borrowings
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: memberId
          in: query
          schema:
            type: integer
        - name: bookId
          in: query
          schema:
            type: integer
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of borrowings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BorrowingListResponse"
    post:
      operationId: createBorrowing
      tags: [borrowings]
      summary: Create a borrowing
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BorrowingInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Borrowing"
  /borrowings/{id}/return:
    post:
      operationId: returnBook
      tags: [borrowings]
      summary: Return a borrowed book
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Book returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Borrowing"
...
components:
  schemas:
...
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
...
    Book:
      type: object
      required: [id, isbn, title, author, stock, status]
      properties:
        id:
          type: integer
        isbn:
          type: string
        title:
          type: string
        author:
          type: string
        publisher:
          type: ["string", "null"]
        year:
          type: ["integer", "null"]
        stock:
          type: integer
        available:
          type: integer
        description:
          type: ["string", "null"]
        pages:
          type: ["integer", "null"]
        categoryId:
          type: ["integer", "null"]
        rackId:
          type: ["integer", "null"]
        coverUrl:
          type: ["string", "null"]
        categoryName:
          type: ["string", "null"]
        rackName:
          type: ["string", "null"]
        borrowCount:
          type: integer
        status:
          type: string
          enum: [available, borrowed]
        createdAt:
          type: ["string", "null"]
    BookListResponse:
      type: object
      required: [data, total]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Book"
        total:
          type: integer
    BookInput:
      type: object
      required: [isbn, title, author]
      properties:
        isbn:
          type: string
        title:
          type: string
        author:
          type: string
        publisher:
          type: string
        year:
          type: integer
        stock:
          type: integer
        description:
          type: string
        pages:
          type: integer
        categoryId:
          type: integer
        rackId:
          type: integer
        coverUrl:
          type: string
    Category:
      type: object
      required: [id, name]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        bookCount:
          type: integer
        createdAt:
          type: ["string", "null"]
    CategoryInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: string
...
    Borrowing:
      type: object
      required: [id, memberId, bookId, borrowDate, dueDate, status]
      properties:
        id:
          type: integer
        borrowCode:
          type: ["string", "null"]
        memberId:
          type: integer
        bookId:
          type: integer
        borrowDate:
          type: string
        dueDate:
          type: string
        returnDate:
          type: ["string", "null"]
        status:
          type: string
          enum: [borrowed, returned, overdue]
        fine:
          type: number
        notes:
          type: ["string", "null"]
        memberName:
          type: ["string", "null"]
        memberNumber:
          type: ["string", "null"]
        bookTitle:
          type: ["string", "null"]
        bookIsbn:
          type: ["string", "null"]
        createdAt:
          type: ["string", "null"]
    BorrowingListResponse:
      type: object
      required: [data, total]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Borrowing"
        total:
          type: integer
    BorrowingInput:
      type: object
      required: [memberId, bookId, dueDate]
      properties:
        memberId:
          type: integer
        bookId:
          type: integer
        dueDate:
          type: string
        notes:
          type: string
...