# lib/db/src/schema/books.ts
import { pgTable, serial, text, integer, timestamp, check, unique } from "drizzle-orm/pg-core";
import { createInsertSchema } from "drizzle-zod";
import { z } from "zod/v4";
import { sql } from "drizzle-orm";

export const booksTable = pgTable("books", {
  id: serial("id").primaryKey(),
  isbn: text("isbn").notNull(),
  title: text("title").notNull(),
  author: text("author").notNull(),
  publisher: text("publisher"),
  year: integer("year"),
  stock: integer("stock").notNull().default(1),
  description: text("description"),
  pages: integer("pages"),
  categoryId: integer("category_id"),
  rackId: integer("rack_id"),
  coverUrl: text("cover_url"),
  createdAt: timestamp("created_at", { withTimezone: true }).notNull().defaultNow(),
  updatedAt: timestamp("updated_at", { withTimezone: true }).notNull().defaultNow().$onUpdate(() => new Date()),
}, (t) => [
  check("stock_non_negative", sql`${t.stock} >= 0`),
]);

export const insertBookSchema = createInsertSchema(booksTable).omit({ id: true, createdAt: true, updatedAt: true });
export type InsertBook = z.infer<typeof insertBookSchema>;
export type Book = typeof booksTable.$inferSelect;